[ALPS05882258] Disp: fix 3 coverity scanned code defect
[Detail]
fix 3 coverity scanned code defect

MTK-Commit-Id: 39642536ac9cd0437285c9742c63bebc347ebfc1

CR-Id: ALPS05882258
Feature: Display Driver (DSI Interface)

+++ drivers/misc/mediatek/video/mt6768/dispsys/ddp_disp_bdg.c
@@ -925,7 +925,7 @@ int bdg_mipi_tx_dphy_clk_setting(enum DISP_BDG_ENUM module,
 			DSI_OUTREGBIT(cmdq, struct MIPI_TX_PHY_SEL0_REG,
 					MIPI_TX_REG->MIPI_TX_PHY_SEL0,
 					MIPI_TX_PHY1_SEL,
-					pad_mapping[swap_base[MIPITX_PHY_LANE_1]]);
+					pad_mapping[swap_base[(unsigned int)MIPITX_PHY_LANE_1]]);
 			DSI_OUTREGBIT(cmdq, struct MIPI_TX_PHY_SEL1_REG,
 					MIPI_TX_REG->MIPI_TX_PHY_SEL1,
 					MIPI_TX_PHY2BC_SEL,
